# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Comment management service for Minisource",
        "title": "Comment Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5010",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/comments/{id}/reactions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reactions"
                ],
                "summary": "Add or update a reaction to a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reaction data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.ReactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reactions"
                ],
                "summary": "Remove a reaction from a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{id}/reactions/me": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reactions"
                ],
                "summary": "Get current user's reaction to a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.UserReactionResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_minisource_comment_internal_models.ReactionRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "type": {
                    "enum": [
                        "like",
                        "dislike",
                        "love",
                        "haha",
                        "wow",
                        "sad",
                        "angry"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.ReactionType"
                        }
                    ]
                }
            }
        },
        "github_com_minisource_comment_internal_models.ReactionType": {
            "type": "string",
            "enum": [
                "like",
                "dislike",
                "love",
                "haha",
                "wow",
                "sad",
                "angry"
            ],
            "x-enum-varnames": [
                "ReactionLike",
                "ReactionDislike",
                "ReactionLove",
                "ReactionHaha",
                "ReactionWow",
                "ReactionSad",
                "ReactionAngry"
            ]
        },
...
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_handler.UserReactionResponse": {
            "type": "object",
            "properties": {
                "comment_id": {
                    "type": "string"
                },
                "has_reacted": {
                    "type": "boolean"
                },
                "reaction_type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}